+++ source/source/gameengine/Device/DEV_InputDevice.h
@@ -39,6 +39,11 @@ protected:
 	std::map<int, SCA_EnumInputs> m_reverseButtonTranslateTable;
 	std::map<int, SCA_EnumInputs> m_reverseWindowTranslateTable;
 
+	/** Teclas seguradas pelo teclado fisico e pelo controle na tela (Web), em separado: soltar o toque nao solta
+	 * a tecla que o teclado ainda segura, e vice-versa. O evento so muda quando o estado combinado muda. */
+	bool m_physicalKeys[MAX_KEYS];
+	bool m_virtualKeys[MAX_KEYS];
+
 public:
 	DEV_InputDevice();
 	virtual ~DEV_InputDevice();
@@ -49,6 +54,8 @@ public:
 	void ConvertMoveEvent(int x, int y);
 	void ConvertWheelEvent(int z);
 	void ConvertEvent(SCA_IInputDevice::SCA_EnumInputs type, int val, unsigned int unicode);
+	/// Le as teclas seguradas pelo controle na tela (Module.rangePad.keys); so age no build Web.
+	void PollVirtualKeys();
 };
 
 #endif  // __DEV_INPUTDEVICE_H__